+++ embeddings/glove.py
@@ -1,7 +1,6 @@
 import argparse
 from collections import Counter
 import logging
-import os
 from pathlib import Path
 import struct
